requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
imbalanced-learn==0.12.4
matplotlib==3.9.2
python-dotenv==1.0.1
googlemaps==4.10.0
geopy==2.4.1
twilio==9.3.0

# Optional but commonly used
requests==2.32.3

